submit/pyproject.toml
[tool.poetry]
name = "submit"
version = "0.1.0"
description = ""
authors = ["Dennis Brakhane <brakhane@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,^3.10"
trio = "0.21.0"
starlette = "^0.21.0"
hypercorn = {extras = ["trio"], version = "^0.14.3"}
python-multipart = "^0.0.5"
starlite = "^1.25.0"
PyNaCl = "^1.5.0"
rich = "^12.6.0"
httpx = "^0.23.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
